requirements.txt
azure-ai-inference
azure-search-documents
python-dotenv
aiohttp
azure-identity